requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1
httpx==0.25.2
openai==1.3.7
anthropic==0.7.8
cohere==4.37
python-dotenv==1.0.0
python-multipart==0.0.6
websockets==12.0
prometheus-client==0.19.0
structlog==23.2.0
celery==5.3.4
kafka-python==2.0.2
numpy==1.24.3
pandas==2.1.4
scikit-learn==1.3.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
tenacity==8.2.3
asyncio-mqtt==0.16.1
